+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added support for the compressor parameter `Boost Amount`
+
 ### Fixed
 
 - The threshold parameter from the deesser plugin is now saved to the preset file